More integration test refactors to speed them up
