+++ web-client/integration-tests/caseJourney.test.js
@@ -54,7 +54,7 @@ import taxPayerSignsOut from './journey/taxpayerSignsOut';
 import taxpayerCancelsCreateCase from './journey/taxpayerCancelsCreateCase';
 import taxpayerChoosesCaseType from './journey/taxpayerChoosesCaseType';
 import taxpayerChoosesProcedureType from './journey/taxpayerChoosesProcedureType';
-import taxpayerCreatesNewCase from './journey/taxpayerCreatesNewCase';
+import taxpayerCreatesNewCaseTestAllOptions from './journey/taxpayerCreatesNewCaseTestAllOptions';
 import taxpayerLogin from './journey/taxpayerLogIn';
 import taxpayerNavigatesToCreateCase from './journey/taxpayerCancelsCreateCase';
 import taxpayerViewsCaseDetail from './journey/taxpayerViewsCaseDetail';
@@ -122,7 +122,7 @@ describe('Case journey', () => {
   taxpayerNavigatesToCreateCase(test);
   taxpayerChoosesProcedureType(test);
   taxpayerChoosesCaseType(test);
-  taxpayerCreatesNewCase(test, fakeFile);
+  taxpayerCreatesNewCaseTestAllOptions(test, fakeFile);
   taxpayerViewsDashboard(test);
   taxpayerViewsCaseDetail(test);
   taxPayerSignsOut(test);
